+++ Classes/DrugDealer.cs
@@ -1,22 +1,21 @@
 ﻿using GTA.Math;
 using Los.Santos.Dope.Wars.Classes.Base;
 
-namespace Los.Santos.Dope.Wars.Classes
+namespace Los.Santos.Dope.Wars.Classes;
+
+/// <summary>
+/// The <see cref="DrugDealer"/> class, inherits from the base class <see cref="Dealer"/>
+/// </summary>
+public class DrugDealer : Dealer
 {
+	#region ctor
 	/// <summary>
-	/// The <see cref="DrugDealer"/> class, inherits from the base class <see cref="Dealer"/>
+	/// The <see cref="DrugDealer"/> standard constructor
 	/// </summary>
-	public class DrugDealer : Dealer
+	/// <param name="position"></param>
+	/// <param name="heading"></param>
+	public DrugDealer(Vector3 position, float heading) : base(position, heading)
 	{
-		#region ctor
-		/// <summary>
-		/// The <see cref="DrugDealer"/> standard constructor
-		/// </summary>
-		/// <param name="position"></param>
-		/// <param name="heading"></param>
-		public DrugDealer(Vector3 position, float heading) : base(position, heading)
-		{
-		}
-		#endregion
 	}
+	#endregion
 }
