+++ Classes/PlayerStash.cs
@@ -1,60 +1,59 @@
 ﻿using GTA;
 using Los.Santos.Dope.Wars.Classes.Base;
-using Los.Santos.Dope.Wars.Contracts;
+using Los.Santos.Dope.Wars.Interfaces;
 
-namespace Los.Santos.Dope.Wars.Classes
+namespace Los.Santos.Dope.Wars.Classes;
+
+/// <summary>
+/// The <see cref="DealerStash"/> class for player types, inherits from <see cref="Stash"/>
+/// </summary>
+public class PlayerStash : Stash, IPlayerStash
 {
+	#region ctor
 	/// <summary>
-	/// The <see cref="DealerStash"/> class for player types, inherits from <see cref="Stash"/>
+	/// The standard constructor for the <see cref="PlayerStash"/> class
 	/// </summary>
-	public class PlayerStash : Stash, IPlayerStash
+	public PlayerStash() { }
+	#endregion
+
+	#region IPlayerStash members
+	/// <inheritdoc/>
+	public void BuyDrug(string drugName, int drugQuantity, int drugPrice)
 	{
-		#region ctor
-		/// <summary>
-		/// The standard constructor for the <see cref="PlayerStash"/> class
-		/// </summary>
-		public PlayerStash() { }
-		#endregion
+		Drug drug = Drugs.Where(x => x.Name.Equals(drugName, StringComparison.Ordinal)).SingleOrDefault();
 
-		#region IPlayerStash members
-		/// <inheritdoc/>
-		public void BuyDrug(string drugName, int drugQuantity, int drugPrice)
-		{
-			Drug drug = Drugs.Where(x => x.Name.Equals(drugName, StringComparison.Ordinal)).SingleOrDefault();
+		if (drug.Quantity.Equals(0))
+			drug.PurchasePrice = drugPrice;
+		else
+			drug.PurchasePrice = ((drug.Quantity * drug.PurchasePrice) + (drugQuantity * drugPrice)) / (drug.Quantity + drugQuantity);
 
-			if (drug.Quantity.Equals(0))
-				drug.PurchasePrice = drugPrice;
-			else
-				drug.PurchasePrice = ((drug.Quantity * drug.PurchasePrice) + (drugQuantity * drugPrice)) / (drug.Quantity + drugQuantity);
+		AddToStash(drugName, drugQuantity);
 
-			AddToStash(drugName, drugQuantity);
-
-			Game.Player.Money -= drugPrice * drugQuantity;
-		}
-		/// <inheritdoc/>
-		public void MoveIntoInventory(string drugName, int drugQuantity, int drugPrice)
-		{
-			Drug drug = Drugs.Where(x => x.Name.Equals(drugName, StringComparison.Ordinal)).SingleOrDefault();
-
-			if (drug.Quantity.Equals(0))
-				drug.PurchasePrice = drugPrice;
-			else
-				drug.PurchasePrice = ((drug.Quantity * drug.PurchasePrice) + (drugQuantity * drugPrice)) / (drug.Quantity + drugQuantity);
-
-			AddToStash(drugName, drugQuantity);
-		}
-		/// <inheritdoc/>
-		public void SellDrug(string drugName, int drugQuantity, int drugPrice)
-		{
-			RemoveFromStash(drugName, drugQuantity);
-
-			Game.Player.Money += drugPrice * drugQuantity;
-		}
-		/// <inheritdoc/>
-		public void TakeFromInventory(string drugName, int drugQuantity)
-		{
-			RemoveFromStash(drugName, drugQuantity);
-		}
-		#endregion
+		Game.Player.Money -= drugPrice * drugQuantity;
 	}
+	/// <inheritdoc/>
+	public void MoveIntoInventory(string drugName, int drugQuantity, int drugPrice)
+	{
+		Drug drug = Drugs.Where(x => x.Name.Equals(drugName, StringComparison.Ordinal)).SingleOrDefault();
+
+		if (drug.Quantity.Equals(0))
+			drug.PurchasePrice = drugPrice;
+		else
+			drug.PurchasePrice = ((drug.Quantity * drug.PurchasePrice) + (drugQuantity * drugPrice)) / (drug.Quantity + drugQuantity);
+
+		AddToStash(drugName, drugQuantity);
+	}
+	/// <inheritdoc/>
+	public void SellDrug(string drugName, int drugQuantity, int drugPrice)
+	{
+		RemoveFromStash(drugName, drugQuantity);
+
+		Game.Player.Money += drugPrice * drugQuantity;
+	}
+	/// <inheritdoc/>
+	public void TakeFromInventory(string drugName, int drugQuantity)
+	{
+		RemoveFromStash(drugName, drugQuantity);
+	}
+	#endregion
 }
