+++ GUI/Elements/DrugListItem.cs
@@ -2,162 +2,161 @@
 using Los.Santos.Dope.Wars.Classes;
 using static Los.Santos.Dope.Wars.Enums;
 
-namespace Los.Santos.Dope.Wars.GUI.Elements
+namespace Los.Santos.Dope.Wars.GUI.Elements;
+
+/// <summary>
+/// The <see cref="DrugListItem"/> class serves as the buy and sell menu list item, abstracts from <see cref="NativeListItem{T}"/>
+/// </summary>
+public class DrugListItem : NativeListItem<int>
 {
+	#region fields
+	private readonly bool _isPlayerItem;
+	#endregion
+
+	#region constructor
 	/// <summary>
-	/// The <see cref="DrugListItem"/> class serves as the buy and sell menu list item, abstracts from <see cref="NativeListItem{T}"/>
+	/// The <see cref="DrugListItem"/> standard constructor
 	/// </summary>
-	public class DrugListItem : NativeListItem<int>
+	/// <param name="drug"></param>
+	/// <param name="isPlayerItem"></param>
+	public DrugListItem(Drug drug, bool isPlayerItem = false) : base($" {drug.Name}", GetIntArrayFromDrugQuantity(drug.Quantity))
 	{
-		#region fields
-		private readonly bool _isPlayerItem;
-		#endregion
-
-		#region constructor
-		/// <summary>
-		/// The <see cref="DrugListItem"/> standard constructor
-		/// </summary>
-		/// <param name="drug"></param>
-		/// <param name="isPlayerItem"></param>
-		public DrugListItem(Drug drug, bool isPlayerItem = false) : base($" {drug.Name}", GetIntArrayFromDrugQuantity(drug.Quantity))
-		{
-			_isPlayerItem = isPlayerItem;
-			if (drug.Quantity.Equals(0))
-				Enabled = false;
-			Tag = drug;
-			SelectedIndex = drug.Quantity;
-			LeftBadgeSet = GetDrugBadgeSet(drug.Name);
-			RefreshDescription(drug, drug.Quantity, isPlayerItem);
-			ItemChanged += OnItemChanged;
-		}
-		#endregion
-
-		#region private methods
-		/// <summary>
-		/// The <see cref="OnItemChanged(object, ItemChangedEventArgs{int})"/> simply refreshes the item description if the item value has changed
-		/// </summary>
-		/// <param name="sender"></param>
-		/// <param name="e"></param>
-		private void OnItemChanged(object sender, ItemChangedEventArgs<int> e)
-		{
-			if (Tag is not Drug drug)
-				return;
-			RefreshDescription(drug, e.Object, _isPlayerItem);
-		}
-
-		/// <summary>
-		/// The <see cref="GetIntArrayFromDrugQuantity(int)"/> returns an array of integers starting from zero ...
-		/// </summary>
-		/// <param name="quantity"></param>
-		/// <returns><see cref="Array"/></returns>
-		private static int[] GetIntArrayFromDrugQuantity(int quantity)
-		{
-			int[] array = new int[quantity + 1];
-			for (int i = 0; i <= quantity; i++)
-				array[i] = i;
-			return array;
-		}
-
-		/// <summary>
-		/// The <see cref="GetDrugBadgeSet(string)"/> method returns a corresponding badgeset
-		/// </summary>
-		/// <param name="drugName"></param>
-		/// <returns><see cref="BadgeSet"/></returns>
-		private static BadgeSet GetDrugBadgeSet(string drugName)
-		{
-			BadgeSet cokeBadgeSet = new("mpinventory", "mp_specitem_coke", "mp_specitem_coke_black");
-			BadgeSet methBadgeSet = new("mpinventory", "mp_specitem_meth", "mp_specitem_meth_black");
-			BadgeSet weedBadgeSet = new("mpinventory", "mp_specitem_weed", "mp_specitem_weed_black");
-			BadgeSet heroBadgeSet = new("mpinventory", "mp_specitem_heroin", "mp_specitem_heroin_black");
-			BadgeSet defaultBadgeSet = new("mpinventory", "mp_specitem_cash", "mp_specitem_cash_black");
-
-			if (DrugType.Cocaine.ToString().Equals(drugName, StringComparison.Ordinal))
-				return cokeBadgeSet;
-			if (DrugType.Heroin.ToString().Equals(drugName, StringComparison.Ordinal))
-				return heroBadgeSet;
-			if (DrugType.Marijuana.ToString().Equals(drugName, StringComparison.Ordinal))
-				return weedBadgeSet;
-			if (DrugType.Hashish.ToString().Equals(drugName, StringComparison.Ordinal))
-				return weedBadgeSet;
-			if (DrugType.Mushrooms.ToString().Equals(drugName, StringComparison.Ordinal))
-				return weedBadgeSet;
-			if (DrugType.Amphetamine.ToString().Equals(drugName, StringComparison.Ordinal))
-				return cokeBadgeSet;
-			if (DrugType.PCP.ToString().Equals(drugName, StringComparison.Ordinal))
-				return cokeBadgeSet;
-			if (DrugType.Methamphetamine.ToString().Equals(drugName, StringComparison.Ordinal))
-				return methBadgeSet;
-			if (DrugType.Ketamine.ToString().Equals(drugName, StringComparison.Ordinal))
-				return cokeBadgeSet;
-			if (DrugType.Mescaline.ToString().Equals(drugName, StringComparison.Ordinal))
-				return methBadgeSet;
-			if (DrugType.Ecstasy.ToString().Equals(drugName, StringComparison.Ordinal))
-				return methBadgeSet;
-			if (DrugType.Acid.ToString().Equals(drugName, StringComparison.Ordinal))
-				return methBadgeSet;
-			if (DrugType.MDMA.ToString().Equals(drugName, StringComparison.Ordinal))
-				return methBadgeSet;
-			if (DrugType.Crack.ToString().Equals(drugName, StringComparison.Ordinal))
-				return heroBadgeSet;
-			if (DrugType.Oxycodone.ToString().Equals(drugName, StringComparison.Ordinal))
-				return methBadgeSet;
-			else return defaultBadgeSet;
-		}
-
-		/// <summary>
-		/// The <see cref="RefreshDescription(Drug, int, bool)"/> method refreshes the item description
-		/// </summary>
-		/// <param name="drug"></param>
-		/// <param name="drugQuantity"></param>
-		/// <param name="isPlayerItem"></param>
-		private void RefreshDescription(Drug drug, int drugQuantity, bool isPlayerItem)
-		{
-			int valueOne = drug.CurrentPrice;
-			int valueTwo = isPlayerItem ? drug.PurchasePrice : drug.AveragePrice;
-
-			string pon = GetGoodOrBadPrice(valueOne, valueTwo, isPlayerItem);
-			string inPercent = GetDifferenceInPercent(valueOne, valueTwo, isPlayerItem);
-
-			Description = isPlayerItem ? $"Purchase price:\t${drug.PurchasePrice}\n" : $"Market price:\t\t${drug.AveragePrice}\n";
-			Description = Description +
-				$"Current price:\t\t{pon}${drug.CurrentPrice} ({inPercent})\n" +
-				$"~w~Purchase price:\t{drugQuantity} x {pon}${drug.CurrentPrice} ~w~= {pon}${drugQuantity * drug.CurrentPrice}";
-		}
-
-		/// <summary>
-		/// Returns price difference in percent, already colored
-		/// </summary>
-		/// <param name="valueOne">Should be the current price</param>
-		/// <param name="valueTwo">Should be the market or purchase price</param>
-		/// <param name="isPlayer"></param>
-		/// <returns><see cref="string"/></returns>
-		private static string GetDifferenceInPercent(int valueOne, int valueTwo, bool isPlayer = false)
-		{
-			double resultValue = (valueOne / (double)valueTwo * 100) - 100;
-
-			if (resultValue > 0)
-				return $"{(isPlayer ? "~g~+" : "~r~+")}{resultValue:n2}%";
-			else if (resultValue < 0)
-				return $"{(isPlayer ? "~r~" : "~g~")}{resultValue:n2}%";
-			else
-				return $"~w~{resultValue:n2}%";
-		}
-
-		/// <summary>
-		/// Returns the color it green, red or white string if it good, bad or neutral
-		/// </summary>
-		/// <param name="valueOne">Should be the current price</param>
-		/// <param name="valueTwo">Should be the market or purchase price</param>
-		/// <param name="isPlayer"></param>
-		/// <returns><see cref="string"/></returns>
-		private static string GetGoodOrBadPrice(int valueOne, int valueTwo, bool isPlayer = false)
-		{
-			if (valueOne > valueTwo)
-				return isPlayer ? "~g~" : "~r~";
-			if (valueOne < valueTwo)
-				return isPlayer ? "~r~" : "~g~";
-			return "~w~";
-		}
-		#endregion
+		_isPlayerItem = isPlayerItem;
+		if (drug.Quantity.Equals(0))
+			Enabled = false;
+		Tag = drug;
+		SelectedIndex = drug.Quantity;
+		LeftBadgeSet = GetDrugBadgeSet(drug.Name);
+		RefreshDescription(drug, drug.Quantity, isPlayerItem);
+		ItemChanged += OnItemChanged;
 	}
+	#endregion
+
+	#region private methods
+	/// <summary>
+	/// The <see cref="OnItemChanged(object, ItemChangedEventArgs{int})"/> simply refreshes the item description if the item value has changed
+	/// </summary>
+	/// <param name="sender"></param>
+	/// <param name="e"></param>
+	private void OnItemChanged(object sender, ItemChangedEventArgs<int> e)
+	{
+		if (Tag is not Drug drug)
+			return;
+		RefreshDescription(drug, e.Object, _isPlayerItem);
+	}
+
+	/// <summary>
+	/// The <see cref="GetIntArrayFromDrugQuantity(int)"/> returns an array of integers starting from zero ...
+	/// </summary>
+	/// <param name="quantity"></param>
+	/// <returns><see cref="Array"/></returns>
+	private static int[] GetIntArrayFromDrugQuantity(int quantity)
+	{
+		int[] array = new int[quantity + 1];
+		for (int i = 0; i <= quantity; i++)
+			array[i] = i;
+		return array;
+	}
+
+	/// <summary>
+	/// The <see cref="GetDrugBadgeSet(string)"/> method returns a corresponding badgeset
+	/// </summary>
+	/// <param name="drugName"></param>
+	/// <returns><see cref="BadgeSet"/></returns>
+	private static BadgeSet GetDrugBadgeSet(string drugName)
+	{
+		BadgeSet cokeBadgeSet = new("mpinventory", "mp_specitem_coke", "mp_specitem_coke_black");
+		BadgeSet methBadgeSet = new("mpinventory", "mp_specitem_meth", "mp_specitem_meth_black");
+		BadgeSet weedBadgeSet = new("mpinventory", "mp_specitem_weed", "mp_specitem_weed_black");
+		BadgeSet heroBadgeSet = new("mpinventory", "mp_specitem_heroin", "mp_specitem_heroin_black");
+		BadgeSet defaultBadgeSet = new("mpinventory", "mp_specitem_cash", "mp_specitem_cash_black");
+
+		if (DrugType.Cocaine.ToString().Equals(drugName, StringComparison.Ordinal))
+			return cokeBadgeSet;
+		if (DrugType.Heroin.ToString().Equals(drugName, StringComparison.Ordinal))
+			return heroBadgeSet;
+		if (DrugType.Marijuana.ToString().Equals(drugName, StringComparison.Ordinal))
+			return weedBadgeSet;
+		if (DrugType.Hashish.ToString().Equals(drugName, StringComparison.Ordinal))
+			return weedBadgeSet;
+		if (DrugType.Mushrooms.ToString().Equals(drugName, StringComparison.Ordinal))
+			return weedBadgeSet;
+		if (DrugType.Amphetamine.ToString().Equals(drugName, StringComparison.Ordinal))
+			return cokeBadgeSet;
+		if (DrugType.PCP.ToString().Equals(drugName, StringComparison.Ordinal))
+			return cokeBadgeSet;
+		if (DrugType.Methamphetamine.ToString().Equals(drugName, StringComparison.Ordinal))
+			return methBadgeSet;
+		if (DrugType.Ketamine.ToString().Equals(drugName, StringComparison.Ordinal))
+			return cokeBadgeSet;
+		if (DrugType.Mescaline.ToString().Equals(drugName, StringComparison.Ordinal))
+			return methBadgeSet;
+		if (DrugType.Ecstasy.ToString().Equals(drugName, StringComparison.Ordinal))
+			return methBadgeSet;
+		if (DrugType.Acid.ToString().Equals(drugName, StringComparison.Ordinal))
+			return methBadgeSet;
+		if (DrugType.MDMA.ToString().Equals(drugName, StringComparison.Ordinal))
+			return methBadgeSet;
+		if (DrugType.Crack.ToString().Equals(drugName, StringComparison.Ordinal))
+			return heroBadgeSet;
+		if (DrugType.Oxycodone.ToString().Equals(drugName, StringComparison.Ordinal))
+			return methBadgeSet;
+		else return defaultBadgeSet;
+	}
+
+	/// <summary>
+	/// The <see cref="RefreshDescription(Drug, int, bool)"/> method refreshes the item description
+	/// </summary>
+	/// <param name="drug"></param>
+	/// <param name="drugQuantity"></param>
+	/// <param name="isPlayerItem"></param>
+	private void RefreshDescription(Drug drug, int drugQuantity, bool isPlayerItem)
+	{
+		int valueOne = drug.CurrentPrice;
+		int valueTwo = isPlayerItem ? drug.PurchasePrice : drug.AveragePrice;
+
+		string pon = GetGoodOrBadPrice(valueOne, valueTwo, isPlayerItem);
+		string inPercent = GetDifferenceInPercent(valueOne, valueTwo, isPlayerItem);
+
+		Description = isPlayerItem ? $"Purchase price:\t${drug.PurchasePrice}\n" : $"Market price:\t\t${drug.AveragePrice}\n";
+		Description = Description +
+			$"Current price:\t\t{pon}${drug.CurrentPrice} ({inPercent})\n" +
+			$"~w~Purchase price:\t{drugQuantity} x {pon}${drug.CurrentPrice} ~w~= {pon}${drugQuantity * drug.CurrentPrice}";
+	}
+
+	/// <summary>
+	/// Returns price difference in percent, already colored
+	/// </summary>
+	/// <param name="valueOne">Should be the current price</param>
+	/// <param name="valueTwo">Should be the market or purchase price</param>
+	/// <param name="isPlayer"></param>
+	/// <returns><see cref="string"/></returns>
+	private static string GetDifferenceInPercent(int valueOne, int valueTwo, bool isPlayer = false)
+	{
+		double resultValue = (valueOne / (double)valueTwo * 100) - 100;
+
+		if (resultValue > 0)
+			return $"{(isPlayer ? "~g~+" : "~r~+")}{resultValue:n2}%";
+		else if (resultValue < 0)
+			return $"{(isPlayer ? "~r~" : "~g~")}{resultValue:n2}%";
+		else
+			return $"~w~{resultValue:n2}%";
+	}
+
+	/// <summary>
+	/// Returns the color it green, red or white string if it good, bad or neutral
+	/// </summary>
+	/// <param name="valueOne">Should be the current price</param>
+	/// <param name="valueTwo">Should be the market or purchase price</param>
+	/// <param name="isPlayer"></param>
+	/// <returns><see cref="string"/></returns>
+	private static string GetGoodOrBadPrice(int valueOne, int valueTwo, bool isPlayer = false)
+	{
+		if (valueOne > valueTwo)
+			return isPlayer ? "~g~" : "~r~";
+		if (valueOne < valueTwo)
+			return isPlayer ? "~r~" : "~g~";
+		return "~w~";
+	}
+	#endregion
 }
